+++ web-common/src/features/dashboards/filters/dimension-filters/constants.ts
